restore old example, move new demo to new file
